Move additional xy debug methods to common file
