+++ packages/charts/src/chart_types/bullet_graph/renderer/canvas/bullet.ts
@@ -11,9 +11,8 @@ import { Color, Colors } from '../../../../common/colors';
 import { Ratio } from '../../../../common/geometry';
 import { cssFontShorthand } from '../../../../common/text_utils';
 import { withContext, clearCanvas } from '../../../../renderers/canvas';
-import { renderDebugRect } from '../../../../renderers/canvas/utils/debug';
+import { renderDebugRect, renderDebugPoint } from '../../../../renderers/canvas/utils/debug';
 import { A11ySettings } from '../../../../state/selectors/get_accessibility_config';
-import { renderDebugPoint } from '../../../xy_chart/renderer/canvas/utils/debug';
 import { ActiveValue } from '../../selectors/get_active_values';
 import { BulletDimensions } from '../../selectors/get_panel_dimensions';
 import { BulletSpec, BulletSubtype } from '../../spec';
