+++ packages/charts/src/chart_types/xy_chart/renderer/canvas/axes/tick_label.ts
@@ -8,9 +8,9 @@
 
 import { AxisProps } from '.';
 import { renderText } from '../../../../../renderers/canvas/primitives/text';
+import { renderDebugRectCenterRotated } from '../../../../../renderers/canvas/utils/debug';
 import { Position } from '../../../../../utils/common';
 import { AxisTick, getTickLabelPosition } from '../../../utils/axis_utils';
-import { renderDebugRectCenterRotated } from '../utils/debug';
 
 const TICK_TO_LABEL_GAP = 2;
